refactor: isolating creation from object creation method
